backend/requirements.txt
flask~=3.1.1
python-dotenv~=1.1.1
flask-sqlalchemy~=3.1.1
pydantic~=2.11.7
argon2-cffi~=25.1.0
psycopg2-binary~=2.9.10
pyjwt~=2.10.1
fastapi-mail~=1.5.2
